refactor: 'lib/api.ts'로 todo api 함수들 이동
