+++ components/Todos/TodoForm.tsx
@@ -3,7 +3,7 @@ import Image from "next/image";
 import Button from "@/components/Button";
 import Input from "@/components/Input";
 import FormControl from "@/components/FormControl";
-import addTodo from "./addTodo";
+import { addTodo } from "@/lib/api";
 import { useMutation, useQueryClient } from "@tanstack/react-query";
 import { Item } from "@/types/todo";
